+++ src/storage/indexmapped.rs
@@ -2,17 +2,17 @@ use scc::{TreeIndex, ebr::Guard};
 
 use crate::{Cell, storage::StorageFor};
 
-use super::OutputType;
+use super::Computation;
 
 /// TreeIndexStorage is backed internally by a `scc::TreeIndex`,
 /// a type optimized for read-heavy workloads. See [scc's documentation](https://docs.rs/scc/2.3.4/scc/#treeindex)
 /// for performance details.
-pub struct TreeIndexStorage<K: OutputType> {
+pub struct TreeIndexStorage<K: Computation> {
     key_to_cell: TreeIndex<K, Cell>,
     cell_to_key: TreeIndex<Cell, (K, Option<K::Output>)>,
 }
 
-impl<K: OutputType> Default for TreeIndexStorage<K> {
+impl<K: Computation> Default for TreeIndexStorage<K> {
     fn default() -> Self {
         Self {
             key_to_cell: Default::default(),
@@ -23,7 +23,7 @@ impl<K: OutputType> Default for TreeIndexStorage<K> {
 
 impl<K> StorageFor<K> for TreeIndexStorage<K>
 where
-    K: Clone + Ord + OutputType + 'static,
+    K: Clone + Ord + Computation + 'static,
     K::Output: Clone + Eq,
 {
     fn get_cell_for_computation(&self, key: &K) -> Option<Cell> {
@@ -35,10 +35,9 @@ where
         self.cell_to_key.insert(cell, (key, None)).ok();
     }
 
-    fn get_input(&self, cell: Cell) -> K {
+    fn try_get_input(&self, cell: Cell) -> Option<K> {
         self.cell_to_key
             .peek_with(&cell, |_, (k, _)| k.clone())
-            .unwrap()
     }
 
     fn get_output(&self, cell: Cell) -> Option<K::Output> {
@@ -86,7 +85,7 @@ where
 
 impl<K> serde::Serialize for TreeIndexStorage<K>
 where
-    K: serde::Serialize + OutputType + Eq + Clone + 'static,
+    K: serde::Serialize + Computation + Eq + Clone + 'static,
     K::Output: serde::Serialize + Clone + 'static,
 {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -107,7 +106,7 @@ where
 
 impl<'de, K> serde::Deserialize<'de> for TreeIndexStorage<K>
 where
-    K: serde::Deserialize<'de> + Ord + OutputType + Clone + 'static,
+    K: serde::Deserialize<'de> + Ord + Computation + Clone + 'static,
     K::Output: serde::Deserialize<'de> + Clone,
 {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>